build various versions of a package

build various versions of a package

Usage

build_pkg(
  .pkgdir = ".",
  types = c("source", "binary"),
  repository = NULL,
  origin = NULL,
  addl_meta = NULL,
  supplement_version = FALSE,
  overwrite = TRUE,
  ...
)

Arguments

.pkgdir

path to package

types

types of package to build

repository

repository name being built for

origin

package source

addl_meta

additional metadata

supplement_version

add additional version info (unix timestamp) to version. TRUE inspects the pkg folder as a git repo, also may provide a character or numeric value to append

overwrite

overwrite fields already present when adding fields

...

parameters to pass to pkgbuild

Details

supplementing version can be done whenever a build occurs that does not correspond to a formal release/tag. This will automatically add information about the git hash (if available), as well as incrememnt the version number with a unix timestamp that corresponds to the last git hash (if present) or the current system time, if git is not present and no version timestamp is provided.